@@ -15,11 +15,11 @@ classifiers =
     Programming Language :: Cython
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
     Development Status :: 4 - Beta
     Intended Audience :: Developers
     License :: OSI Approved :: GNU General Public License v3 (GPLv3)
@@ -31,7 +31,7 @@ include_package_data = True
 package_dir =
     =src
 packages = find_namespace:
-python_requires = >=3.6, <4.0
+python_requires = >=3.7, <4.0
 setup_requires =
   wheel
   setuptools>=50.3.2
